requirements.txt
wheel 
Pillow
djangorestframework
djangorestframework-simplejwt

drf-yasg
psycopg2-binary
python-decouple
python-slugify



#
# Windows
#py -m venv venv
#venv\Scripts\activate

# Mac / Linux
#python3 -m venv venv
#. venv/bin/activate

#pip install -r requirements.txt

#pip install ______